# cmd/api/internal/swaggerdoc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is all apis for the trip-itinary application.",
        "title": "Trip-Itinary API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/day_sessions/{id}/stop": {
            "get": {
                "description": "Lists all the planstops under a plan version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plan"
                ],
                "summary": "List Plan Stops",
                "parameters": [
                    {
                        "type": "string",
                        "description": "daysession ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/plans_application_dto.PlanStopDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.apiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.apiError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates planstopd for each planversion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plan"
                ],
                "summary": "Creates PlanStop",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PlanVersion ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "PlanStop details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.CreatePlanStopRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/plans_application_dto.PlanStopDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.apiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_module_plans_infrastructure_driving_rest.apiError"
                        }
                    }
                }
            }
        },
...
        "/day-sessions/assistant-suggestions/{id}": {
            "put": {
                "description": "Edit assistant suggestion for the specified day session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assistant_Suggestions"
                ],
                "summary": "Edit assistant suggestion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Day Session ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Edit Assistant suggestion details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_assistant_suggestions_infrastructure_driving_rest.EditAssistantSuggestionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/assistant_suggestions_application_dto.AssistantSuggestionsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_module_assistant_suggestions_infrastructure_driving_rest.apiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_module_assistant_suggestions_infrastructure_driving_rest.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_module_assistant_suggestions_infrastructure_driving_rest.apiError"
                        }
                    }
                }
            }
        },
        "/day-sessions/{id}/assistant-suggestions": {
            "get": {
                "description": "Get all the assistant suggestions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assistant_Suggestions"
                ],
                "summary": "Get assistant suggestions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Day Session ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/assistant_suggestions_application_dto.AssistantSuggestionsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_module_assistant_suggestions_infrastructure_driving_rest.apiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_module_assistant_suggestions_infrastructure_driving_rest.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_module_assistant_suggestions_infrastructure_driving_rest.apiError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new assistant suggestion for the specified day session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assistant_Suggestions"
                ],
                "summary": "Create an assistant suggestion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Day Session ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Assistant suggestion details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_assistant_suggestions_infrastructure_driving_rest.CreateAssistantSuggestionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/assistant_suggestions_application_dto.AssistantSuggestionsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_module_assistant_suggestions_infrastructure_driving_rest.apiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_module_assistant_suggestions_infrastructure_driving_rest.apiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_module_assistant_suggestions_infrastructure_driving_rest.apiError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "assistant_suggestions_application_dto.AssistantSuggestionsDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "daysessionID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_module_assistant_suggestions_infrastructure_driving_rest.CreateAssistantSuggestionRequest": {
            "type": "object",
            "properties": {
                "day_session_id": {
                    "description": "Day session id",
                    "type": "string"
                },
                "message": {
                    "description": "Message by the assistant",
                    "type": "string"
                },
                "status": {
                    "description": "Status (pending, accepted, snoozed)",
                    "type": "string"
                }
            }
        },
        "internal_module_assistant_suggestions_infrastructure_driving_rest.EditAssistantSuggestionRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Edited Message",
                    "type": "string"
                },
                "status": {
                    "description": "Edited Status",
                    "type": "string"
                }
            }
        },
        "internal_module_assistant_suggestions_infrastructure_driving_rest.apiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        },
                        "requestID": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "internal_module_plans_infrastructure_driving_rest.CreatePlanStopRequest": {
            "type": "object",
            "properties": {
                "busyrisklabel": {
                    "description": "Busy Risk Label",
                    "type": "string"
                },
                "categorylabel": {
                    "description": "Category of the Place",
                    "type": "string"
                },
                "imageurl": {
                    "description": "Image URL of the Place",
                    "type": "string"
                },
                "plannedarrival": {
                    "description": "Arrival Time",
                    "type": "string"
                },
                "planneddeparture": {
                    "description": "Departure Time",
                    "type": "string"
                },
                "position": {
                    "description": "Starting Place number",
                    "type": "integer"
                },
                "stayminutes": {
                    "description": "Spendable time in the Place",
                    "type": "integer"
                },
                "title": {
                    "description": "Place name",
                    "type": "string"
                },
                "travelminutes": {
                    "description": "Travel Time",
                    "type": "integer"
                }
            }
        },
...
        "internal_module_plans_infrastructure_driving_rest.apiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        },
                        "requestID": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "plans_application_dto.PlanStopDTO": {
            "type": "object",
            "properties": {
                "busyRiskLabel": {
                    "type": "string"
                },
                "categoryLabel": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "imageURL": {
                    "type": "string"
                },
                "plannedArrival": {
                    "type": "string"
                },
                "plannedDeparture": {
                    "type": "string"
                },
                "planversionID": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "stayMinutes": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "travelMinutes": {
                    "type": "integer"
                }
            }
        },
...
    }
}